+++ Pandora/Paradigm/Structure/Ability/Insertable.hs
@@ -1,0 +1,4 @@
+module Pandora.Paradigm.Structure.Ability.Insertable where
+
+class Insertable t where
+	insert :: a -> t a -> t a