+++ Pandora/Pattern/Functor.hs
@@ -1,4 +1,4 @@
-module Pandora.Pattern.Functor (module Exports) where
+module Pandora.Pattern.Functor (module Exports, (<*+>), (<**+>)) where
 
 import Pandora.Pattern.Functor.Bivariant as Exports
 import Pandora.Pattern.Functor.Divariant as Exports
@@ -19,3 +19,16 @@ import Pandora.Pattern.Functor.Alternative as Exports
 import Pandora.Pattern.Functor.Invariant as Exports
 import Pandora.Pattern.Functor.Contravariant as Exports
 import Pandora.Pattern.Functor.Covariant as Exports
+
+import Pandora.Core.Functor (type (:.), type (:=))
+
+(<*+>) :: (Applicative t, Alternative u) => t :. u := a -> t :. u := a -> t :. u := a
+x <*+> y = (<+>) <$> x <*> y
+
+(<**+>) :: (Applicative t, Applicative u, Alternative v)
+	=> t :. u :. v := a -> t :. u :. v := a -> t :. u :. v := a
+x <**+> y = (<+>) <$$> x <**> y
+
+(<***+>) :: (Applicative t, Applicative u, Applicative v, Alternative w)
+	=> t :. u :. v :. w := a -> t :. u :. v :. w := a -> t :. u :. v :. w := a
+x <***+> y = (<+>) <$$$> x <***> y