+++ Pandora/Paradigm/Structure.hs
@@ -1,6 +1,7 @@
 module Pandora.Paradigm.Structure (module Exports) where
 
 import Pandora.Paradigm.Structure.Ability as Exports
+import Pandora.Paradigm.Structure.Interface as Exports
 import Pandora.Paradigm.Structure.Rose as Exports
 import Pandora.Paradigm.Structure.Splay as Exports
 import Pandora.Paradigm.Structure.Binary as Exports