+++ Pandora/Paradigm/Structure/Interface.hs
@@ -1,0 +1,3 @@
+module Pandora.Paradigm.Structure.Interface (module Exports) where
+
+import Pandora.Paradigm.Structure.Interface.Set as Exports