+++ CHANGELOG.md
@@ -254,3 +254,4 @@
 * Add experimental `Monotonic` typeclass
 * Implement delete method for Stack
 * Define `Comprehension` wrapper for data structures that can behave like list comprehensions
+* Define experimental `Insertable` typeclass